Add option [connections].current_speed_time_interval_msecs to segget.conf
CURRENT_SPEED_TIME_INTERVAL_MSECS
segget transfers may have bursty nature of their traffic. Therefore, while
measuring current speed, segget actually calculates average speed during
current_speed_time_interval_msecs time interval, defined in milliseconds.
Min limit:100
Max limit: 60000
Default:
current_speed_time_interval_msecs=1000

+++ segget/config.cpp
@@ -1,3 +1,29 @@
+/*
+* Copyright (C) 2010 Robin H.Johnson, Ovechko Kostyantyn <fastinetserver@gmail.com>.
+*
+* Project: IDFetch.
+* Developer: Ovechko Kostyantyn Olexandrovich (Kharkiv State Technical University of Construction and Architecture, Ukraine).
+* Mentor: Robin H. Johnson (Gentoo Linux: Developer, Trustee & Infrastructure Lead).
+* Mentoring organization: Gentoo Linux.
+* Sponsored by GSOC 2010.
+*
+* This file is part of Segget.
+*
+* Segget is free software; you can redistribute it and/or
+* modify it under the terms of the GNU Lesser General Public
+* License as published by the Free Software Foundation; either
+* version 2.1 of the License, or (at your option) any later version.
+*
+* Segget is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+* Lesser General Public License for more details.
+*
+* You should have received a copy of the GNU Lesser General Public
+* License along with Segget; if not, write to the Free Software
+* Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+*/
+
 #ifndef __CONFIG_H__
 #define __CONFIG_H__
 
@@ -16,7 +42,7 @@ public:
 	ConfigFile(std::string const& configFile);
 
 	int set(string &dst,string const& section, string const& entry) const;
-	int set(uint &dst,string const& section, string const& entry, uint const& min_limit, uint const& max_limit) const;
+	int set(ulong &dst,string const& section, string const& entry, uint const& min_limit, uint const& max_limit) const;
 	int set(bool &dst,string const& section, string const& entry) const;
 };
 
@@ -92,7 +118,7 @@ int ConfigFile::set(string &dst, string const& section, string const& entry) con
 		return 0;
 	}
 }
-int ConfigFile::set(uint &dst, string const& section, string const& entry, uint const& min_limit, uint const& max_limit) const {
+int ConfigFile::set(ulong &dst, string const& section, string const& entry, uint const& min_limit, uint const& max_limit) const {
 	uint return_value;
 	map<string,string>::const_iterator ci = content_.find(section + '/' + entry);
